@@ -1,60 +1,30 @@
 {
-  "name": "ww-weekly-schedule",
+  "name": "tags-input-component",
   "version": "1.0.0",
-  "description": "An interactive weekly schedule component for WeWeb that allows users to plan their week with different activity types while enforcing a maximum study hours limit.",
-  "author": {
-    "name": "Your Name",
-    "email": "your.email@example.com"
-  },
+  "description": "Componente de entrada de tags para WeWeb que permite adicionar e remover tags facilmente",
+  "main": "src/wwElement.vue",
   "license": "MIT",
-  "keywords": [
-    "weweb",
-    "component",
-    "schedule",
-    "timetable",
-    "weekly",
-    "planner",
-    "study",
-    "hours",
-    "calendar",
-    "time-management"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/yourusername/ww-weekly-schedule.git"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "main": "ww-weekly-schedule.umd.js",
   "scripts": {
     "build": "weweb build",
     "serve": "weweb serve"
   },
   "dependencies": {
-    "vue": "^2.6.14"
+    "vue": "^3.2.37"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^5.0.8",
-    "@vue/cli-plugin-eslint": "^5.0.8",
-    "@vue/cli-service": "^5.0.8",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^8.0.3",
-    "sass": "^1.43.4",
-    "sass-loader": "^10.2.0",
-    "vue-template-compiler": "^2.6.14"
+    "@weweb/cli": "latest"
   },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ],
   "weweb": {
-    "componentType": "wwObject",
-    "name": "Weekly Schedule",
-    "category": "interaction",
-    "keepSvgAttributes": false,
-    "wewebVersion": ">=1.9.0"
-  }
+    "configFile": "ww-config.js",
+    "componentPath": "src/wwElement.vue"
+  },
+  "keywords": [
+    "weweb",
+    "component",
+    "tags",
+    "input",
+    "disciplines",
+    "multi-select",
+    "chips"
+  ]
 }
